cheack leaaks

+++ libft/ft_calloc.c
@@ -16,7 +16,7 @@ void	*ft_calloc(size_t count, size_t size)
 {
 	void	*ptr;
 
-	if ((size != 0) && (count > (UINT32_MAX / size)))
+	if ((size != 0) && (count > (SIZE_MAX / size)))
 		return (NULL);
 	ptr = malloc(count * size);
 	if (ptr != NULL)
